backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
psycopg2-binary
pycryptodome
google-auth
python-dotenv
requests